+++ src/controls/number.ts
@@ -380,3 +380,5 @@ export default class WUPNumberControl<
 }
 
 customElements.define(tagName, WUPNumberControl);
+
+// todo Type Long Text, Press clear, Press Ctrl+Z